Acces check update && Product ready for test

+++ src/main/java/com/example/RESTftulSN/repositories/UsersRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UsersRepository extends JpaRepository<Users,Integer> {
+public interface UsersRepository extends JpaRepository<Users,Long> {
     Optional<Users> findByUsername(String username);
     Boolean existsByUsername(String username);
     Boolean existsByEmail(String email);
